+++ filament/backend/src/vulkan/VulkanSwapChain.h
@@ -20,23 +20,30 @@
 #include "VulkanContext.h"
 #include "VulkanDriver.h"
 
+#include <memory>
+
 #include <utils/FixedCapacityVector.h>
 
 namespace filament {
 namespace backend {
 
 struct VulkanSwapChain : public HwSwapChain {
-    VulkanSwapChain(VulkanContext& context, VkSurfaceKHR vksurface);
-    VulkanSwapChain(VulkanContext& context, uint32_t width, uint32_t height);
+    VulkanSwapChain(VulkanContext& context, VulkanStagePool& stagePool, VkSurfaceKHR vksurface);
+
+    // Headless constructor.
+    VulkanSwapChain(VulkanContext& context, VulkanStagePool& stagePool, uint32_t width, uint32_t height);
 
     bool acquire();
-    void create();
+    void create(VulkanStagePool& stagePool);
     void destroy();
     void makePresentable();
     bool hasResized() const;
-    VulkanAttachment& getColor() { return color[currentSwapIndex]; }
+    VulkanAttachment getColorAttachment() const; // TODO: remove
+    VulkanAttachment getDepthAttachment() const; // TODO: remove
+    VulkanTexture& getColorTexture() const;
 
-    VulkanContext& context;
+    // TODO: privatize more fields
+
     VkSurfaceKHR surface = {};
     VkSwapchainKHR swapchain = {};
     VkSurfaceFormatKHR surfaceFormat = {};
@@ -45,23 +52,26 @@ struct VulkanSwapChain : public HwSwapChain {
     VkQueue headlessQueue = {};
     uint32_t currentSwapIndex = {};
 
-    // Color attachments are swapped, but depth is not. Typically there are 2 or 3 color attachments
-    // in a swap chain.
-    utils::FixedCapacityVector<VulkanAttachment> color;
-    VulkanAttachment depth = {};
-
     // This is signaled when vkAcquireNextImageKHR succeeds, and is waited on by the first
     // submission.
-    VkSemaphore imageAvailable = {};
+    VkSemaphore imageAvailable = VK_NULL_HANDLE;
 
     // This is true after the swap chain image has been acquired, but before it has been presented.
     bool acquired = false;
 
     bool suboptimal = false;
     bool firstRenderPass = false;
+
+private:
+    VulkanContext& mContext;
+
+    // Color attachments are swapped, but depth is not. Typically there are 2 or 3 color attachments
+    // in a swap chain.
+    utils::FixedCapacityVector<std::unique_ptr<VulkanTexture>> mColor;
+    std::unique_ptr<VulkanTexture> mDepth;
 };
 
 } // namespace filament
 } // namespace backend
 
-#endif // TNT_FILAMENT_DRIVER_VULKANTEXTURE_H
+#endif // TNT_FILAMENT_DRIVER_VULKANSWAPCHAIN_H