+++ filament/include/filament/RenderableManager.h
@@ -20,6 +20,7 @@
 #include <filament/Box.h>
 #include <filament/FilamentAPI.h>
 #include <filament/MaterialEnums.h>
+#include <filament/MorphTargetBuffer.h>
 
 #include <backend/DriverEnums.h>
 
@@ -41,7 +42,6 @@ class Engine;
 class IndexBuffer;
 class Material;
 class MaterialInstance;
-class MorphTargetBuffer;
 class Renderer;
 class SkinningBuffer;
 class VertexBuffer;
@@ -485,12 +485,14 @@ public:
     void setMorphTargetBufferAt(Instance instance, uint8_t level, size_t primitiveIndex,
             MorphTargetBuffer* morphTargetBuffer, size_t offset, size_t count);
 
-    void setMorphTargetBufferAt(Instance instance, uint8_t level, size_t primitiveIndex,
-            MorphTargetBuffer* morphTargetBuffer, size_t count); //!< \overload
+    /**
+     * Utility method to change a MorphTargetBuffer to the given primitive
+     */
+    inline void setMorphTargetBufferAt(Instance instance, uint8_t level, size_t primitiveIndex,
+            MorphTargetBuffer* morphTargetBuffer);
 
     /**
-     * Gets t
-     * number of morphing in the given entity.
+     * Gets the number of morphing in the given entity.
      */
     size_t getMorphTargetCount(Instance instance) const noexcept;
 
@@ -595,6 +597,12 @@ public:
     static Box computeAABB(VECTOR const* vertices, INDEX const* indices, size_t count,
             size_t stride = sizeof(VECTOR)) noexcept;
 };
+
+void RenderableManager::setMorphTargetBufferAt(Instance instance, uint8_t level, size_t primitiveIndex,
+        MorphTargetBuffer* morphTargetBuffer) {
+    setMorphTargetBufferAt(instance, level, primitiveIndex, morphTargetBuffer, 0,
+            morphTargetBuffer->getVertexCount());
+}
 
 template<typename VECTOR, typename INDEX, typename, typename>
 Box RenderableManager::computeAABB(VECTOR const* vertices, INDEX const* indices, size_t count,