+++ libs/filabridge/include/filament/MaterialEnums.h
@@ -27,7 +27,7 @@
 namespace filament {
 
 // update this when a new version of filament wouldn't work with older materials
-static constexpr size_t MATERIAL_VERSION = 19;
+static constexpr size_t MATERIAL_VERSION = 20;
 
 /**
  * Supported shading models
@@ -230,6 +230,18 @@ enum class Property : uint8_t {
 
     // when adding new Properties, make sure to update MATERIAL_PROPERTIES_COUNT
 };
+
+enum class UserVariantFilterBit : uint32_t {
+    DIRECTIONAL_LIGHTING        = 0x01,
+    DYNAMIC_LIGHTING            = 0x02,
+    SHADOW_RECEIVER             = 0x04,
+    SKINNING                    = 0x08,
+    FOG                         = 0x10,
+    VSM                         = 0x20,
+    SSR                         = 0x40,
+};
+
+using UserVariantFilterMask = uint32_t;
 
 } // namespace filament